+++ bluetooth_low_energy_platform_interface/lib/src/peripheral_manager.dart
@@ -5,31 +5,37 @@ import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 import 'advertisement.dart';
 import 'bluetooth_low_energy_manager.dart';
 import 'central.dart';
+import 'connection_state.dart';
 import 'event_args.dart';
 import 'gatt.dart';
 
 /// The central connection state cahnged event arguments.
-final class CentralConnectionStateChangedEventArgs
-    extends ConnectionStateChangedEventArgs {
+final class CentralConnectionStateChangedEventArgs extends EventArgs {
   /// The central which connection state changed.
   final Central central;
+
+  /// The connection state.
+  final ConnectionState state;
 
   /// Constructs a [CentralConnectionStateChangedEventArgs].
   CentralConnectionStateChangedEventArgs(
     this.central,
-    super.state,
+    this.state,
   );
 }
 
 /// The central MTU changed event arguments.
-final class CentralMTUChangedEventArgs extends MTUChangedEventArgs {
+final class CentralMTUChangedEventArgs extends EventArgs {
   /// The central which MTU changed.
   final Central central;
+
+  /// The MTU.
+  final int mtu;
 
   /// Constructs a [CentralMTUChangedEventArgs].
   CentralMTUChangedEventArgs(
     this.central,
-    super.mtu,
+    this.mtu,
   );
 }
 
@@ -38,12 +44,16 @@ final class GATTCharacteristicReadRequestedEventArgs extends EventArgs {
   /// The central which read this characteristic.
   final Central central;
 
+  /// The characteristic to read the value of.
+  final GATTCharacteristic characteristic;
+
   /// The read request.
-  final GATTCharacteristicReadRequest request;
+  final GATTReadRequest request;
 
   /// Constructs a [GATTCharacteristicReadRequestedEventArgs].
   GATTCharacteristicReadRequestedEventArgs(
     this.central,
+    this.characteristic,
     this.request,
   );
 }
@@ -53,12 +63,16 @@ final class GATTCharacteristicWriteRequestedEventArgs extends EventArgs {
   /// The central which wrote this characteristic.
   final Central central;
 
+  /// The characteristic to write the value of.
+  final GATTCharacteristic characteristic;
+
   /// The write request.
-  final GATTCharacteristicWriteRequest request;
+  final GATTWriteRequest request;
 
   /// Constructs a [GATTCharacteristicWriteRequestedEventArgs].
   GATTCharacteristicWriteRequestedEventArgs(
     this.central,
+    this.characteristic,
     this.request,
   );
 }
@@ -84,30 +98,38 @@ final class GATTCharacteristicNotifyStateChangedEventArgs extends EventArgs {
 
 /// The GATT descriptor read requested event arguments.
 final class GATTDescriptorReadRequestedEventArgs extends EventArgs {
-  /// The central which read this characteristic.
+  /// The central which read this descriptor.
   final Central central;
 
+  /// The descriptor to read the value of.
+  final GATTDescriptor descriptor;
+
   /// The read request.
-  final GATTDescriptorReadRequest request;
+  final GATTReadRequest request;
 
   /// Constructs a [GATTDescriptorReadRequestedEventArgs].
   GATTDescriptorReadRequestedEventArgs(
     this.central,
+    this.descriptor,
     this.request,
   );
 }
 
 /// The GATT descriptor write requested event arguments.
 final class GATTDescriptorWriteRequestedEventArgs extends EventArgs {
-  /// The central which wrote this characteristic.
+  /// The central which wrote this descriptor.
   final Central central;
 
+  /// The descriptor to write the value of.
+  final GATTDescriptor descriptor;
+
   /// The write request.
-  final GATTDescriptorWriteRequest request;
+  final GATTWriteRequest request;
 
   /// Constructs a [GATTDescriptorWriteRequestedEventArgs].
   GATTDescriptorWriteRequestedEventArgs(
     this.central,
+    this.descriptor,
     this.request,
   );
 }
@@ -186,65 +208,80 @@ abstract interface class PeripheralManager
 
   /// Responds to a read request from a connected central.
   Future<void> respondCharacteristicReadRequestWithValue(
-    GATTCharacteristicReadRequest request, {
+    Central central, {
+    required GATTCharacteristic characteristic,
+    required GATTReadRequest request,
     required Uint8List value,
   });
 
   /// Responds to a read request from a connected central.
   Future<void> respondCharacteristicReadRequestWithError(
-    GATTCharacteristicReadRequest request, {
+    Central central, {
+    required GATTCharacteristic characteristic,
+    required GATTReadRequest request,
     required GATTError error,
   });
 
   /// Responds to a write request from a connected central.
   Future<void> respondCharacteristicWriteRequest(
-    GATTCharacteristicWriteRequest request,
-  );
+    Central central, {
+    required GATTCharacteristic characteristic,
+    required GATTWriteRequest request,
+  });
 
   /// Responds to a write request from a connected central.
   Future<void> respondCharacteristicWriteRequestWithError(
-    GATTCharacteristicWriteRequest request, {
+    Central central, {
+    required GATTCharacteristic characteristic,
+    required GATTWriteRequest request,
     required GATTError error,
   });
 
   /// Send an updated characteristic value to one or more subscribed centrals,
   /// using a notification or indication.
   ///
+  /// [central] A central (represented by CBCentral objects) that have subscribed
+  /// to receive updates of the characteristic’s value. The manager ignores any
+  /// centrals that haven’t subscribed to the characteristic’s value.
+  ///
   /// [characteristic] The characteristic whose value has changed.
   ///
   /// [value] The characteristic value you want to send via a notification or
   /// indication.
-  ///
-  /// [central] A list of centrals (represented by CBCentral objects) that have
-  /// subscribed to receive updates of the characteristic’s value. If nil, the
-  /// manager updates all subscribed centrals. The manager ignores any centrals
-  /// that haven’t subscribed to the characteristic’s value.
   Future<void> notifyCharacteristic(
-    GATTCharacteristic characteristic, {
+    Central central, {
+    required GATTCharacteristic characteristic,
     required Uint8List value,
-    Central? central,
   });
 
   /// Responds to a read request from a connected central.
   Future<void> respondDescriptorReadRequestWithValue(
-    GATTDescriptorReadRequest request, {
+    Central central, {
+    required GATTDescriptor descriptor,
+    required GATTReadRequest request,
     required Uint8List value,
   });
 
   /// Responds to a read request from a connected central.
   Future<void> respondDescriptorReadRequestWithError(
-    GATTDescriptorReadRequest request, {
+    Central central, {
+    required GATTDescriptor descriptor,
+    required GATTReadRequest request,
     required GATTError error,
   });
 
   /// Responds to a write request from a connected central.
   Future<void> respondDescriptorWriteRequest(
-    GATTDescriptorWriteRequest request,
-  );
+    Central central, {
+    required GATTDescriptor descriptor,
+    required GATTWriteRequest request,
+  });
 
   /// Responds to a write request from a connected central.
   Future<void> respondDescriptorWriteRequestWithError(
-    GATTDescriptorWriteRequest request, {
+    Central central, {
+    required GATTDescriptor descriptor,
+    required GATTWriteRequest request,
     required GATTError error,
   });
 }